+++ api/models/profile.py
@@ -60,40 +60,3 @@ class Profile(models.Model):
     # icon = models.ImageField(default="default.jpg", upload_to="profile_pics") NEEDS PILLOW
 
 
-class Ticket(models.Model):
-    """
-    The Ticket class for Sluggo. This will store all information associated with a specific ticket.
-
-    The class contains:
-        owner: A foreign key to a specific user that authored the ticket. Allows for them to edit the ticket.
-        assigned_user: A foreign key that refers to a specific user that the ticket is assigned to.
-        title: A char field for the title of the ticket (currently limited to 100 characters and is a required field).
-        description: A multiline text field that will store the longer form explanation of the ticket.
-        created: A datetime field that will record when a ticket has been made.
-        started: A ticket can be made before anyone actually starts it, so the started field must be seperate. (Also datetime)
-        completed: A field to record when a ticket has been finished. (Datetime as well)
-        due_date: The due date for the ticket, a date field that will keep track of when things are due.
-    """
-
-    owner = models.ForeignKey(
-        settings.AUTH_USER_MODEL,
-        related_name="created_tickets",
-        on_delete=models.CASCADE,
-    )
-    assigned_user = models.ForeignKey(
-        settings.AUTH_USER_MODEL,
-        related_name="assigned_tickets",
-        on_delete=models.CASCADE,
-    )
-    title = models.CharField(max_length=100, blank=False)
-    description = models.TextField(blank=True)
-    created = models.DateTimeField(auto_now_add=True)
-    started = models.DateTimeField(blank=True, null=True)
-    completed = models.DateTimeField(blank=True, null=True)
-    due_date = models.DateField(blank=True, null=True)
-
-    class Meta:
-        ordering = ["id"]
-
-    def __str__(self):
-        return f"Ticket: {self.title}"
